Downloads/backend_gala_og/requirements.txt
google-auth==2.23.0
pyjwt==2.8.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
